reply and tag removed from annotation

+++ swe573_2022_repo/annotation/views/delete_annotation.py
@@ -6,16 +6,18 @@ from django.contrib.auth.models import User
 from django.http import JsonResponse
 from django.core import serializers
 from annotation.models import  AnnotationModel
-
+import json
 
 @login_required(login_url='/')
 @csrf_exempt
 def delete_annotation(request):
     if  request.method == "POST":
-
-        ann_id= request.POST.get('annotationid')
-        ann = AnnotationModel.objects.get(annotation_id=ann_id)
-        ann.delete()
-        return JsonResponse({"data": 'success'}, status=200)
+        annotation= json.loads(request.POST.get('annotation'))
+        ann = AnnotationModel.objects.get(annotation__0__icontains=annotation["id"])
+        annotationJson = json.loads(ann.annotation)
+        if annotationJson["body"][0]["creator"]["name"] == request.user.username:
+            ann.delete()
+            return JsonResponse({"data": 'success'}, status=200)
+        return JsonResponse({"error": "Unauthorized"}, status = 403)
 
     return JsonResponse({"error": "Please try again later"}, status=400)
